Unify TcpWriter and UdpWriter
